requirements.txt
# nvcc --version -> pip install torch==1.7.1+cu101 -f https://download.pytorch.org/whl/torch_stable.html
# pip install pipreqs -> pipreqs to generate requirements.txt
matplotlib==3.7.2
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
scikit_learn==1.3.0
seaborn==0.12.2
torch==1.8.0+cu101
torchvision==0.9.0+cu101
